Seperate PtsQueue and PtsHistory
